requirements.txt
Django
Pillow
WeasyPrint
django-crispy-forms
crispy-bootstrap5
